Update Partay.java
This class is the thing I need to majorly work on. So far I tried to get the entire sorting of the tables in the constructor of this class. Although the code does break when you try and run the program. I need to figure out more logical parts of the main for loop in here that attempts to slot people in. I kinda wrote it on a whim so its very janky at the moment.